funcionality for upload img and show them on principal homeview
