+++ src/App.js
@@ -2,7 +2,6 @@ import React from "react";
 import Login from "./components/LoginUser";
 import Register from "./components/RegisterUser";
 import LeerPost from "./components/LeerPost";
-import SubirPost from "./components/SubirPost";
 import { BrowserRouter as Router, Switch, Route } from "react-router-dom";
 import './App.css'
 
@@ -12,13 +11,13 @@ function App() {
       <div className="container">
         <Switch>
           <Route path="/" exact>
-            <Login />
+            <Login />          
           </Route>
           <Route path="/signin">
             <Register />
           </Route>
-          <Route path="/userhome" >
-            Inicio...
+          <Route path="/userhome" >         
+            <LeerPost />
           </Route>
         </Switch>
       </div>
