+++ src/Components/LeerPost.jsx
@@ -3,6 +3,8 @@ import { firebase } from "../config/firebase";
 
 function LeerPost() {
   const [posts, setPosts] = React.useState([]);
+  const [post, setPost] = React.useState("");
+  const [postImg, setPostImg] = React.useState(false);
 
   React.useEffect(() => {
     const obtenerDatos = async () => {
@@ -23,13 +25,51 @@ function LeerPost() {
     obtenerDatos();
   }, []);
 
+  const adding = async (e) => {
+    e.preventDefault();
+    if (!post.trim()) {
+      console.log("ingrese texto");
+      return;
+    }
+    console.log(post);
+
+    try {
+      const db = firebase.firestore();
+      const newPost = {
+        text: post,
+        img: postImg
+      };
+
+      const data = await db.collection("posts").add(newPost);
+      setPost("");
+      setPosts([...posts, { id: data.id, ...newPost }]);
+    } catch (error) {
+      console.log(error);
+    }
+  };
+
+  const upImg = async (e) => {
+
+    try {
+      const storageRef = firebase.storage().ref();
+      const imgFile = await storageRef.child('pictures/' + e.name).put(e);
+      const imgUrl = await imgFile.ref.getDownloadURL();
+      
+      setPostImg(imgUrl)
+
+    } catch (error) {
+      console.log(error);
+    }
+    
+  }  
+
   const eliminar = async (id) => {
     try {
       const db = firebase.firestore();
       await db.collection("posts").doc(id).delete();
-
       const actualPost = posts.filter((e) => e.id !== id);
       setPosts(actualPost);
+
     } catch (error) {
       console.log(error);
     }
@@ -39,14 +79,42 @@ function LeerPost() {
     <div className="container">
       <div>
         {posts.map((item) => {
-          return (
-            <div>
-              <p key={item.id}>{item.text}</p>
-              <button onClick={() => eliminar(item.id)}>Eliminar</button>
-            </div>
-          );
-        })}
+          if (item.img !== false){ 
+            return (
+              <div key={item.id}>
+                <div>
+                  <p>{item.text}</p>              
+                  <img src={item.img} alt="" />
+                  <button onClick={() => eliminar(item.id)}>Eliminar</button>
+                </div>
+              </div>
+            );
+          } else {
+            return (
+              <div key={item.id}>
+                <p>{item.text}</p>              
+                <button onClick={() => eliminar(item.id)}>Eliminar</button>
+              </div>
+            )
+          }
+        }
+        )}
+
       </div>
+      <h3>Nuevo Post</h3>
+      <form onSubmit={adding}>
+        <input
+          type="text"
+          placeholder="Escribe un nuevo post"
+          onChange={(e) => setPost(e.target.value)}
+          value={post}
+        ></input>
+        <button type="submit">Agregar</button>
+        <input
+        type="file"
+        onChange={e => upImg(e.target.files[0])}
+        />
+      </form>
     </div>
   );
 }
